doctors and teams

+++ src/main/kotlin/com/example/demo/model/MedicalTeam.kt
@@ -1,20 +1,21 @@
 package com.example.demo.model
 
-import java.time.LocalDate
-import javax.persistence.*
+import jakarta.persistence.Entity
+import jakarta.persistence.GeneratedValue
+import jakarta.persistence.GenerationType
+import jakarta.persistence.Id
 
 @Entity
 class MedicalTeam {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    var id: Int? = null
+    var id: Long? = null
     var name: String? = null
 
     constructor() {}
 
-    constructor(id: Int?, name: String?) {
+    constructor(id: Long?, name: String?) {
         this.id = id
-        this.name=name
+        this.name = name
     }
-
 }
